+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   headers.
 - **Breaking**: The upper limit for minimum edge length (`Agedgeinfo_t.minlen`)
   has been expanded from `USHRT_MAX` to `INT_MAX`. #2413
+- **Breaking**: The libcdt macros `DTTREEMATCH` and `DTTREESEARCH` have been
+  removed.
 
 ### Fixed